apps/recommender/requirements.txt
pymongo==4.10.1
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
